Tidy up middlewares
* Fix response body unmarshaling error due to type
