+++ src/api/rest/server/server.go
@@ -16,7 +16,6 @@ package server
 
 import (
 	"context"
-	"encoding/json"
 
 	// "log"
 	"os/signal"
@@ -95,105 +94,11 @@ func RunServer(port string) {
 	// limit the application to 20 requests/sec using the default in-memory store
 	e.Use(middleware.RateLimiter(middleware.NewRateLimiterMemoryStore(20)))
 
-	// Customized middleware for request logging
-	e.Use(func(next echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) error {
-			// log.Debug().Msg("Start - Request ID middleware")
+	// Custom middleware for RequestID and RequestDetails
+	e.Use(middlewares.RequestIdAndDetailsIssuer)
 
-			// Make X-Request-Id visible to all handlers
-			c.Response().Header().Set("Access-Control-Expose-Headers", echo.HeaderXRequestID)
-
-			// Get or generate Request ID
-			reqID := c.Request().Header.Get(echo.HeaderXRequestID)
-			if reqID == "" {
-				reqID = fmt.Sprintf("%d", time.Now().UnixNano())
-			}
-
-			// Set Request on the context
-			c.Set("RequestID", reqID)
-
-			log.Debug().Msgf("(Request ID middleware) Request ID: %s", reqID)
-			if _, ok := common.RequestMap.Load(reqID); ok {
-				return fmt.Errorf("the X-Request-Id is already in use")
-			}
-
-			// Set "X-Request-Id" in response header
-			c.Response().Header().Set(echo.HeaderXRequestID, reqID)
-
-			details := common.RequestDetails{
-				StartTime:   time.Now(),
-				Status:      "Handling",
-				RequestInfo: common.ExtractRequestInfo(c.Request()),
-			}
-			common.RequestMap.Store(reqID, details)
-
-			// log.Debug().Msg("End - Request ID middleware")
-
-			return next(c)
-		}
-	})
-
-	e.Use(middleware.BodyDumpWithConfig(middleware.BodyDumpConfig{
-		Skipper: func(c echo.Context) bool {
-			if c.Path() == "/tumblebug/api" {
-				return true
-			}
-			return false
-		},
-		Handler: func(c echo.Context, reqBody, resBody []byte) {
-			// log.Debug().Msg("Start - BodyDump() middleware")
-
-			// Get the request ID
-			reqID := c.Get("RequestID").(string)
-			log.Debug().Msgf("(BodyDump middleware) Request ID: %s", reqID)
-
-			// Get the content type
-			contentType := c.Response().Header().Get(echo.HeaderContentType)
-			log.Trace().Msgf("contentType: %s", contentType)
-
-			// Dump the response body if content type is "application/json" or "application/json; charset=UTF-8"
-			if contentType == echo.MIMEApplicationJSONCharsetUTF8 || contentType == echo.MIMEApplicationJSON {
-				// Load or check the request by ID
-				if v, ok := common.RequestMap.Load(reqID); ok {
-					log.Trace().Msg("OK, common.RequestMap.Load(reqID)")
-					details := v.(common.RequestDetails)
-					details.EndTime = time.Now()
-
-					// Set "X-Request-Id" in response header
-					c.Response().Header().Set(echo.HeaderXRequestID, reqID)
-
-					// Unmarshal the response body
-					var resMap map[string]interface{}
-					err := json.Unmarshal(resBody, &resMap)
-					if err != nil {
-						log.Error().Err(err).Msgf("Error while unmarshaling response body: %s", err)
-						log.Debug().Msgf("Response body: %s", string(reqBody))
-						log.Debug().Msgf("Response body: %s", string(resBody))
-					}
-
-					// 1XX: Information responses
-					// 2XX: Successful responses (200 OK, 201 Created, 202 Accepted, 204 No Content)
-					// 3XX: Redirection messages
-					// 4XX: Client error responses (400 Bad Request, 401 Unauthorized, 404 Not Found, 408 Request Timeout)
-					// 5XX: Server error responses (500 Internal Server Error, 501 Not Implemented, 503 Service Unavailable)
-					if c.Response().Status >= 400 && c.Response().Status <= 599 {
-						log.Trace().Msg("Error, c.Response().Status")
-						details.Status = "Error"
-						details.ErrorResponse = resMap["message"].(string)
-					} else {
-						log.Trace().Msg("Not error, c.Response().Status")
-						details.Status = "Success"
-						details.ResponseData = resMap
-					}
-
-					// Store details of the request
-					common.RequestMap.Store(reqID, details)
-				}
-			}
-
-			// log.Debug().Msg("Start - BodyDump() middleware")
-		},
-	}))
+	// Custom middleware for ResponseBodyDump
+	e.Use(middlewares.ResponseBodyDump())
 
 	e.HideBanner = true
 	//e.colorer.Printf(banner, e.colorer.Red("v"+Version), e.colorer.Blue(website))
